last truncate
