+++ src/test/java/org/kohsuke/github/GHGistTest.java
@@ -142,20 +142,4 @@ public class GHGistTest extends AbstractGitHubWireMockTest {
         assertThat(f.getContent(), containsString("### Keybase proof"));
     }
 
-    @Test
-    public void truncateTest() throws Exception {
-        char[] data = new char[1000000];
-        String strLess1MB = new String(data);
-        String strGreater1MB = strLess1MB + "abcde";
-        // CRUD operation
-        GHGist gist = gitHub.createGist()
-                .public_(false)
-                .description("Test Truncate")
-                .file("greater1MB.txt", strGreater1MB)
-                .create();
-
-        // test truncate and content value
-        assertThat(gist.getFile("greater1MB.txt").getContent(), equalTo(strLess1MB));
-    }
-
 }
